Active Links
We want to display a link as the active link the user is looking at.
We want that when a user clicks on a link in the left pane that link becomes the active link and display in red.

We swapped out <Link> for <NavLink>
We changed the style prop from a simple object to a function that return an objec
We changed the color of our link by looking at the isActive value that NavLink passed to our styling function

closes #18